Added ColonySidebar
